Function pallet_asset_rate::pallet::dispatchables::update

pub fn update<T: Config>(asset_kind: Box<T::AssetKind>, rate: FixedU128)

Update the conversion rate to native balance for the given asset.

Complexity

O(1)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::update.